year_III/Kwiatek_JNP3/server/requirements.txt
django == 1.11
djangorestframework == 3.7.3
django-appconf == 1.0.2
psycopg2 == 2.7.3.2
aiocoap == 0.3
gunicorn == 19.7.1
django-cors-headers == 2.1.0
django-filter == 1.1.0
enum34
